Merge GITKit based code with Firebase Auth base.
<h1>Rationale:</h1>

The most recent GameChat code base has been built on the Google Identity Toolkit (GITKit) base.  The GITKit is giving way to Firebase authentication.  This commit provides a merged base from which to build out the Firebase auth approach for GameChat.  As such, it is a big and important commit.