First and last take at a Rooms feature...
<h1>Rationale:</h1>

This commit was intended to be a first pass at using the Chat screen to show rooms fragment as a "master" and the chat fragment as the "details".  As this resulted in two masters and two details the whole concept was dropped.  Instead the navigation drawer will be used to select among rooms and experiences (the new name for activities/games).

This commit also fixes a large number of warnings and greases the skids to allow easier and more effective testing in the face of Firebird sign in.

EventBus has been introduced in this commit to permit more loose coupling between software components.  It is not quite totally wired up yet.

The Chat and Game fragments have been moved to "feature" packages out of functional package.  In general feature packages will be preferred going forward with functional packaging used to provide a home for common components used across more than one feature.

Lastly, the min SDK is now set to 18 to allow UIAutomator to be used.  This is still up for debate and may be changed back to 15 if an easier sign in mechanism can be put in place.  One that does not require UIAutomator to pick a sign in choice.