Chess Testing
With a couple of minor bug fixes and testing to check them, and other aspects of the Chess game feature, we are wrapping up on the big 3 games that we wanted to include. Because we now have a substantial number of testing classes, a new sub-package of the test class has been created for all tests that are evaluating any game features.

# Changed Files

## Java Files

### ChessFragment
* Renamed the castling booleans so that they now say “rook” instead of “pawn” (e.g., **“mBlueQueenSideRookHasMoved”** instead of **“mBlueQueenSidePawnHasMoved”**.
* Now correctly restarts the game instead of always making it purple’s turn when a new game is created.

### ChessPiece
* Renamed a few variables so they better describe what they do.

## Test Files

### CheckersTest
* Moved to the new **game** test package.

### GameTest
* Moved to the new **game** test package.

### TTTFragmentTest —> TTTTest
* Renamed
* Moved to the new **game** test package.

# New Files

## Test Files

### ChessTest
* The New Test Class covers…
* the movement ranges of each piece, including pawns moving two spaces from their starting position.
* castling on both sides, and for both players
* starting a new chess game.