Fix missing FAB issue with game settings fragment.
<h1>Rationale:</h1>

The game settings fragment would return to the no games fragment correctly on a back press but the FAB was not being shown.  Debugging revealed that the issue was that the FAB visibility was turned off when starting the settings fragment but was not turned back on when leaving the settings fragment.  Digging deeper, the mechanism used to handle the back press is probably overkill.  The recommended approach is to have the activity capture and handle the back press.  The preferred solution for GameChat is to let the main activity capture the back press and propagate it to other modules via the app event bus. The first module dealing with the back press turns off subsequent event propagation (unless that might not be the right thing to do at the moment).  This commit adopts the more app-centric approach reather than a per-module approach.

This issue also drove home the notion that registering the app event bus during resume and unregistering it during pause is a very good convention because that is when the fragment or activity is running in the foreground, exactly when events should be processed.  Events needing to be processed when the activity or fragment is not running in the foreground should be handled as part of a "service".  This revelation means the Chat world event model needs to be revisited yet again.

And one last observation.  In the game panel, the default "home" base is either the no games fragment or the with games fragment.  This is a bug waiting to land.

<h1>File changes:</h1>

modified:   app/src/main/java/com/pajato/android/gamechat/event/AppEventManager.java

- cancel(), unregisterAll(): new methods providing a richer event bus mechanism.

modified:   app/src/main/java/com/pajato/android/gamechat/game/BaseGameFragment.java

- Summary: make the app event bus follow the recommended conventions for games.
- onAttach(): remove the app event bus registration.
- onDetach(): remove the app event bus de-registration.
- onResume(): add the app event bus registration.
- onPause(): add the app event bus de-registration.

modified:   app/src/main/java/com/pajato/android/gamechat/game/SettingsFragment.java

- onBackPressed(): new method implementing the preferred approach and making the FAB visible.
- onViewCreated(): remove the stale approach to handling the back button.

modified:   app/src/main/java/com/pajato/android/gamechat/main/MainActivity.java

- onBackPressed(): overload to handle the app event post by invoking the system action; handle the system callback by posting and app event.

modified:   app/src/main/java/com/pajato/android/gamechat/main/NavigationManager.java

- onBackPressed(): deal with the app event to detect and close an open nav drawer.

new file:   app/src/main/java/com/pajato/android/gamechat/event/BackPressEvent.java

- Add a simple pojo to define a back press handler.