Force orientation based on device type.
<h1>Rationale:</h1>

GameChat on a smartphone is less than an enjoyably experience when in landscape. The experience is almost as annoying on a tablet in portrait mode.  Therefore this commit forces the orientation on each device to it's natural preference.

<h1>File changes:</h1>

modified:   app/src/main/java/com/pajato/android/gamechat/main/PaneManager.java

- init(): force the orientation based on the device type.