Use back to get to the "home" screen in a game.
<h1>Rationale:</h1>

If the User presses the back key while playing a game, the response should be to return to the top level (no games or list of games.)  This commit provides that feature.  It also ensures that the FAB is turned off on the game pane.

<h1>File changes:</h1>

modified:   app/src/main/java/com/pajato/android/gamechat/game/LocalTTTFragment.java

- onBackPressed(): new subsciption to get the back pressed event, turn on the FAB and return to the "home" game screen.
- onInitialize(): turn the FAB off.