Adding Functionality to the Hamburger Menu
# Rationale

We currently have no way to access the group list or create a new group with the app in its current form. As a sort of band-aid, I have added two functions to the hamburger menu, which is now actually functioning. We can now use it to access the list of groups, which will allow us to make new groups and rooms, and we can always access the Me Room from it, since the Me Group is intentionally hidden from the user.

# Changed Files

## Java Files

### chat.ChatManager
* Two new short methods that facilitate direct movement to the places we'd like to go to directly. One navigates straight to the Me Room, and one navigates straight to the Group List.

### database.GroupManager
* We now no display the Me Group in the ShowGroupListFragment. This is done by not adding it to the group list data whenever the Group Manager is queried about group list data.

### main.MainActivity
* Actually initialize the NavigationManager.
* Handle functionality for our two navigation drawer items. Functionality for other navigation drawer items will come at a later time.

### main.NavigationManager
* Set up the Main Activity to also be the listener for the Navigation Drawer's Footer view, instead of just the movable drawer itself.

## XML Files

### menu/drawer_body
* Added our two new menu items.
* Fixed the dummy backup behind the footer at the end of the menu, so we can properly see everything we want to see in the scrolling part of the drawer.

### values/strings
* Added string resources for our new navigation drawer menu items.