Refactor database list item getters.
<h1>Rationale:</h1>

This commit continues the practice of making small edits (no functional changes) that enhance the app in some way.  In this case there are methods in the database managers that access list item data for list adapters.  Like the setWatcher() methods, these mthods have been given a normalized name, getListItemData().

Also in this push, the basic ToolbarManager class has been included.  It is not referenced but does reveal a design plan.