Update the Background for the Sign In Screen
# Rationale
As the previous image used for the sign in screen dominated the size of
the apk (approximately a third of the space used was one image) it was
decided that an alternate strategy ought to be used instead. So, rather
than an image, the choice was to use an XML gradient drawable, saving
space while keeping up some newer

# Changed Files

## XML Files

### values/colors.xml
* Added a new color, **colorPrimaryLight** which is used for our new
gradient and was specifically chosen to fit the current palette while
being conducive to laying under black text.

### values/styles.xml
* Updated the style for the sign in screen to specify the new background
gradient.

# New Files

## XML Files

### drawable/background_gradient.xml
* The xml file that specifies the background gradient object.

# Deleted Files

## Image Files

### signin_background.png
* As the sign in background was oversized, it has been removed.